a
This is a work in progress fixed version of the zombie survival gamemode created by Necrossin and Deluvas in 2012.  This version starts off where the Mr. Green community left off mid-2015.  Core gameplay hasn't been altered, only minor bugfixes and some code has been brought up to date.  The main goal from this point on is to tackle the horribly unbalanced gameplay and optimize as much code as possible.